Also update unsigned 64-bit integers

+++ src/v1_6/messages/change_availability.rs
@@ -4,7 +4,7 @@ use crate::v1_6::types::{AvailabilityStatus, AvailabilityType};
 #[serde(rename_all = "camelCase")]
 pub struct ChangeAvailabilityRequest {
     /// Required. The id of the connector for which availability needs to change. Id '0' (zero) is used if the availability of the Charge Point and all its connectors needs to change.
-    pub connector_id: u64,
+    pub connector_id: u32,
     /// Required. This contains the type of availability change that the Charge Point should perform.
     #[serde(rename = "type")]
     pub kind: AvailabilityType,
